(CMAKE) adding cmake option for not compiling everything related to LEGACY (i.e SPIDER 1.0) runtime.
