+++ libspider/graphs-tools/transformation/optims/helper/partialSingleRate.cpp
@@ -32,6 +32,9 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL license and that you accept its terms.
  */
+
+#ifndef _NO_BUILD_LEGACY_RT
+
 /* === Include(s) === */
 
 #include <graphs-tools/transformation/optims/helper/partialSingleRate.h>
@@ -98,3 +101,4 @@ void spider::partialSingleRateTransformation(pisdf::Graph *graph,
         }
     }
 }
+#endif
