+++ libspider/graphs-tools/transformation/optims/helper/patternOptimizer.cpp
@@ -32,6 +32,9 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL license and that you accept its terms.
  */
+
+#ifndef _NO_BUILD_LEGACY_RT
+
 /* === Include(s) === */
 
 #include <graphs-tools/transformation/optims/helper/patternOptimizer.h>
@@ -111,3 +114,4 @@ bool spider::optims::reduceFFJJWorker(pisdf::VertexType type,
     }
     return verticesToOptimize.empty();
 }
+#endif
