+++ libspider/graphs-tools/transformation/optims/helper/unitaryOptimizer.cpp
@@ -32,6 +32,9 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL license and that you accept its terms.
  */
+
+#ifndef _NO_BUILD_LEGACY_RT
+
 /* === Include(s) === */
 
 #include <graphs/pisdf/Vertex.h>
@@ -71,3 +74,4 @@ bool spider::optims::optimizeUnitaryVertex(pisdf::Vertex *vertex) {
     }
     return false;
 }
+#endif
