+++ libspider/graphs-tools/transformation/optims/optimizations.cpp
@@ -32,6 +32,9 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL license and that you accept its terms.
  */
+
+#ifndef _NO_BUILD_LEGACY_RT
+
 /* === Include(s) === */
 
 #include <graphs-tools/transformation/optims/optimizations.h>
@@ -431,3 +434,4 @@ bool spider::optims::reduceUnitaryRateActors(const pisdf::Graph *graph) {
     }
     return optimized;
 }
+#endif
