+++ libspider/graphs-tools/transformation/srdag/SingleRateTransformer.cpp
@@ -32,6 +32,9 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL license and that you accept its terms.
  */
+
+#ifndef _NO_BUILD_LEGACY_RT
+
 /* === Include(s) === */
 
 #include <graphs-tools/transformation/srdag/SingleRateTransformer.h>
@@ -550,3 +553,4 @@ void spider::srdag::SingleRateTransformer::populateFromDelayVertex(vector<Transf
     /* == Remove the Edge == */
     srdag_->removeEdge(edge);
 }
+#endif
