+++ libspider/graphs-tools/transformation/srdag/visitors/SRDAGCopyVertexVisitor.cpp
@@ -32,6 +32,8 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL license and that you accept its terms.
  */
+#ifndef _NO_BUILD_LEGACY_RT
+
 /* === Include(s) === */
 
 #include <graphs-tools/transformation/srdag/visitors/SRDAGCopyVertexVisitor.h>
@@ -118,4 +120,4 @@ void spider::srdag::SRDAGCopyVertexVisitor::makeClone(pisdf::Vertex *vertex) {
     }
     ix_ = (srdag_->vertexCount() - 1) - (vertex->repetitionValue() - 1);
 }
-
+#endif
