+++ libspider/graphs-tools/transformation/srdag/visitors/SRDAGCopyVertexVisitor.h
@@ -35,6 +35,8 @@
 #ifndef SPIDER2_SRDAGCOPYVERTEXVISITOR_H
 #define SPIDER2_SRDAGCOPYVERTEXVISITOR_H
 
+#ifndef _NO_BUILD_LEGACY_RT
+
 /* === Include(s) === */
 
 #include <graphs-tools/transformation/srdag/TransfoJob.h>
@@ -65,5 +67,5 @@ namespace spider {
         };
     }
 }
-
+#endif
 #endif //SPIDER2_SRDAGCOPYVERTEXVISITOR_H
