+++ libspider/runtime/algorithm/JITMSRuntime.cpp
@@ -32,7 +32,10 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL license and that you accept its terms.
  */
+
 /* === Include(s) === */
+
+#ifndef _NO_BUILD_LEGACY_RT
 
 #include <runtime/algorithm/JITMSRuntime.h>
 #include <runtime/runner/RTRunner.h>
@@ -252,3 +255,5 @@ void spider::JITMSRuntime::transformDynamicJobs(vector<srdag::TransfoJob> &stati
     /* == Swap vectors == */
     dynamicJobStack.swap(tempJobStack);
 }
+
+#endif
