+++ libspider/scheduling/scheduler/GreedyScheduler.cpp
@@ -32,6 +32,8 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL license and that you accept its terms.
  */
+#ifndef _NO_BUILD_LEGACY_RT
+
 /* === Include(s) === */
 
 #include <scheduling/scheduler/GreedyScheduler.h>
@@ -170,3 +172,4 @@ void spider::sched::GreedyScheduler::removeNonExecutableVertices() {
         }
     }
 }
+#endif
