+++ libspider/scheduling/scheduler/ListScheduler.cpp
@@ -32,6 +32,7 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL license and that you accept its terms.
  */
+#ifndef _NO_BUILD_LEGACY_RT
 
 /* === Include(s) === */
 
@@ -195,3 +196,4 @@ size_t spider::sched::ListScheduler::countNonSchedulableTasks() {
     }
     return static_cast<size_t>(std::distance(sortedTaskVector_.rbegin(), it));
 }
+#endif
